extracted template renderer in separate class
